cli: extract shared OAuth callbacks helper used by init and provider
The OAuth callback construction in src/cli/init.ts (buildOAuthCallbacks)
and src/cli/provider.ts (the inline callbacks in runOAuthLogin) was
~95% duplicated after PR #252 added onManualCodeInput in both places.
Extract to src/cli/oauth-callbacks.ts and re-import from both callsites.

The new helper is a pure function: takes providerName, returns the
OAuthCallbacks shape that src/init/oauth-login.ts forwards to pi-ai.
No behavioral change — the rendered note() box, the onPrompt fallback,
and the onManualCodeInput race prompt are byte-identical to PR #252.

Adds src/cli/oauth-callbacks.test.ts with a minimal mutation-check
guard: asserts every callback the OAuth runner forwards is provided
(including onManualCodeInput). Removing onManualCodeInput from the
helper fails the test, which now covers BOTH CLI entry points via
the single shared definition — addressing the goal-verification gap
flagged in the post-implementation review of PR #252 (the prior tests
only exercised oauth-login.ts wiring, not the CLI callsites).
